Refactor all exercises from previous homework to be solved in single methods that return output specific to exercise requirement and have arguments specific to input.

+++ src/main/java/extraHomework1/PublicMethods.java
@@ -9,7 +9,7 @@ public class PublicMethods {
     }
 
 
-    public static float inchConverter(float a) {
+    public static double inchConverter(float a) {
 
         return (a * 0.0254);
     }
